requirements.txt
aiohttp==3.9.3
aiokafka==0.10.0
aiomultiprocess==0.9.0
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
billiard==4.2.0
black==24.1.1
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
dateparser==1.2.0
exceptiongroup==1.2.0
fastapi==0.109.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
frozenlist==1.4.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
kombu==5.3.5
locust==2.21.0
Mako==1.3.2
MarkupSafe==2.1.4
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nats-py==2.6.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.4
pyzmq==25.1.2
redis==5.0.1
regex==2023.12.25
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
